1.8: UI improvements, can trim single file
If you choose a single file in the browser, instead of giving you an error,
it will adjust the command to skip the filtergraph configuration and simply
trim the file.

UI improvements include a better progress bar indicator (shows 100% if
complete, instead of sometimes 99.97% or something). It also visually
looks different - it's animated while encoding, and solid while done or
canceled. Canceled is a bright red.  Also, for canceled and completed
states, they show up as a bootstrap label instead of just the progress bar
so it's obvious what is currently running and what isn't.